requirements.txt
google-generativeai == 0.3.1
tkinter
mistune == 3.0.2
